requirements.txt
PyYAML>=6.0
numpy>=1.23
matplotlib>=3.6
imageio>=2.31
pillow>=9.5

